backend/requirements.txt
# requirements.txt

alembic>=1.11.0
email-validator>=1.1
factory-boy>=3.2.0
fastapi>=0.100.0
freezegun>=1.1.0
httpx>=0.23.0
passlib[bcrypt]>=1.7.4
psycopg2-binary>=2.9.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pydantic[email]>=2.0.0
pytest>=7.0.0
pytest-cov>=3.0.0
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.5
sqlalchemy>=2.0.0
uvicorn[standard]>=0.23.0
